requirements.txt
ffmpeg-python==0.2.0
future==0.18.2
paho-mqtt==1.5.1
pydub==0.25.1
